add optimize .decrypt, add 2 retries on flush broken fields in profileFex, set timeouts for each field separately
